+++ src/ansys/dpf/core/animator.py
@@ -9,7 +9,8 @@ import numpy as np
 from typing import Union, Sequence
 
 import ansys.dpf.core as core
-from ansys.dpf.core.plotter import _sort_supported_kwargs, _PyVistaPlotter
+from ansys.dpf.core.helpers.utils import _sort_supported_kwargs
+from ansys.dpf.core.plotter import _PyVistaPlotter
 
 
 class _InternalAnimatorFactory:
