+++ src/ansys/dpf/core/plotter.py
@@ -19,9 +19,9 @@ from ansys import dpf
 from ansys.dpf import core
 from ansys.dpf.core.common import locations, DefinitionLabels
 from ansys.dpf.core.common import shell_layers as eshell_layers
-from ansys.dpf.core.helpers import _sort_supported_kwargs
+from ansys.dpf.core.helpers.streamlines import _sort_supported_kwargs
 from ansys.dpf.core import errors as dpf_errors
-from ansys.dpf.core.helpers import compute_streamlines
+from ansys.dpf.core.helpers.streamlines import compute_streamlines
 from ansys.dpf.core.nodes import Node, Nodes
 
 if TYPE_CHECKING:  # pragma: no cover
